### 2020-10-30 #2802 @whedon
[PRE REVIEW]: netrd: A library for network reconstruction and graph distances
**Submitting author:** @sdmccabe (<a href="http://orcid.org/0000-0002-7180-145X">Stefan McCabe</a>)
**Repository:** <a href="https://github.com/netsiphd/netrd" target ="_blank">https://github.com/netsiphd/netrd</a>
**Version:** v0.2.2
