### 2020-10-30 #2802 @arfon
:wave: @sdmccabe, thanks for your submission to JOSS. Could you please make sure you have a section named _Statement of need_ near the top of your paper with a description of the software designed for a generalist audience?
