### 2020-10-30 #2802 @arfon
> 👋 @sdmccabe, thanks for your submission to JOSS. Could you please make sure you have a section named Statement of need near the top of your paper with a description of the software designed for a generalist audience?

Edit: Ignore me. I see the title of your first section. The paper is good as-is.
