### 2020-11-02 6149f97
Replace Travis with Github CI
See #312.
