+++ .travis.yml
@@ -1,56 +1,0 @@
-language: python
-matrix:
-    include:
-        - os: linux
-          sudo: required
-          dist: xenial
-          python: "3.6"
-
-        - os: linux
-          sudo: required
-          dist: xenial
-          python: "3.7"
-
-        - os: linux
-          sudo: required
-          dist: xenial
-          python: "3.8"
-
-        - os: osx
-          language: generic
-          env: PYTHON_TRAVIS_VERSION="3.6"
-
-        - os: osx
-          language: generic
-          env: PYTHON_TRAVIS_VERSION="3.7"
-
-        - os: osx
-          language: generic
-          env: PYTHON_TRAVIS_VERSION="3.8"
-
-before_install:
-  - |
-    if [ "$TRAVIS_OS_NAME" == "osx" ]; then
-       brew cask install miniconda
-    else
-       wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
-       bash miniconda.sh -b -p $HOME/miniconda
-       source "$HOME/miniconda/etc/profile.d/conda.sh"
-       hash -r
-    fi
-    conda config --set always_yes yes --set changeps1 no
-    conda update -q conda
-    source $(conda info --root)/etc/profile.d/conda.sh
-    source .bashrc
-    conda info -a
-install:
-  - conda create -q -n test-env python=$PYTHON_TRAVIS_VERSION black pytest
-  - conda activate test-env
-  - pip install -r requirements.txt
-  - pip install .
-
-before_script: cd tests/
-script:
-  - pytest
-  - black --skip-string-normalization --check ../netrd
-  - black --skip-string-normalization --check ../tests
