+++ .github/python-ci.yml
@@ -1,0 +1,30 @@
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: [3.6, 3.7, 3.8]
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install black pytest
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+    - name: Lint with black
+      run: |
+        cd tests/
+        black --skip-string-normalization --check ../netrd
+        black --skip-string-normalization --check ../tests
+    - name: Test with pytest
+      run: |
+        pytest
