### 2020-11-02 e4dd7a8
fix syntax error in CI workflow
