+++ .github/workflows/python-ci.yml
@@ -14,7 +14,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v2
       with:
-        python-version: [3.6, 3.7, 3.8]
+        python-version: 3.8
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
