+++ .github/workflows/python-ci.yml
@@ -8,13 +8,15 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
+    strategy:
+      python-version: [3.6, 3.7, 3.8]
 
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
-        python-version: 3.8
+        python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
