+++ README.md
@@ -1,8 +1,6 @@
 [![PyPI version](https://badge.fury.io/py/netrd.svg)](https://badge.fury.io/py/netrd)
 [![ReadTheDocs](https://img.shields.io/readthedocs/netrd.svg)](
     https://netrd.readthedocs.io)
-[![Travis](https://img.shields.io/travis/netsiphd/netrd.svg)](
-    https://travis-ci.org/netsiphd/netrd)
 ![CI](https://github.com/netsiphd/netrd/workflows/build/badge.svg)
 
 # netrd: A library for network {reconstruction, distances, dynamics}
