### 2020-11-02 9c60380
correct typo in ci.yml
