+++ .github/workflows/python-ci.yml
@@ -11,7 +11,8 @@ jobs:
 
     runs-on: ubuntu-latest
     strategy:
-      python-version: [3.6, 3.7, 3.8]
+      matrix:
+        python-version: [3.6, 3.7, 3.8]
 
     steps:
     - uses: actions/checkout@v2
