### 2020-11-02 ed29641
add macOS to matrix
