### 2020-11-02 95bbb38
Complete migration to Github Actions for CI (#315)
* fix syntax error in CI workflow

* pip install netrd

* change working directory timing

* Add python-version matrix

* update README badge

* remove Travis badge

* correct typo in ci.yml

* add macOS to matrix

* add windows to matrix

* remove windows from matrix

Co-authored-by: Stefan McCabe <stefanmccabe@gmail.com>
