+++ .github/workflows/python-ci.yml
@@ -1,3 +1,5 @@
+name: build
+
 on:
   push:
     branches: [ master ]
@@ -7,24 +9,29 @@ on:
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [macos-latest, ubuntu-latest]
+        python-version: [3.6, 3.7, 3.8]
 
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
-        python-version: [3.6, 3.7, 3.8]
+        python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install black pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install -e .
     - name: Lint with black
       run: |
-        cd tests/
-        black --skip-string-normalization --check ../netrd
-        black --skip-string-normalization --check ../tests
+        black --skip-string-normalization --check netrd
+        black --skip-string-normalization --check tests
     - name: Test with pytest
       run: |
+        cd tests/
         pytest
