### 2020-11-06 d552d38
Add Python 3.9 support to CI
This _should_ work now that Numpy/Scipy have pushed 3.9 wheels.
