### 2020-11-06 9bac711
Add Python 3.9 support to CI (#317)
This _should_ work now that Numpy/Scipy have pushed 3.9 wheels.
