+++ netrd/distance/netsimile.py
@@ -92,7 +92,6 @@ def feature_extraction(G):
     node_degree_dict = dict(G.degree())
     node_clustering_dict = dict(nx.clustering(G))
     egonets = [nx.ego_graph(G, n) for n in node_list]
-    egonet_dict = {node: egonet for node, egonet in zip(node_list, egonets)}
 
     # node degrees
     degs = [node_degree_dict[n] for n in node_list]
