+++ netrd/reconstruction/correlation_matrix.py
@@ -8,7 +8,6 @@ Submitted as part of the 2019 NetSI Collabathon
 """
 from .base import BaseReconstructor
 import numpy as np
-import networkx as nx
 from ..utilities import create_graph, threshold
 
 
