### 2020-12-17 fc807bb
Clean up unused imports (#319)
See #318. This gets almost all the LGTM issues.
