+++ netrd/utilities/read.py
@@ -10,7 +10,6 @@ Submitted as part of the 2019 NetSI Collabathon.
 
 """
 import numpy as np
-import networkx as nx
 
 
 def read_time_series(filename, delimiter=','):
