+++ .github/workflows/python-ci.yml
@@ -31,6 +31,10 @@ jobs:
       run: |
         black --skip-string-normalization --check netrd
         black --skip-string-normalization --check tests
+    - name: Check for unused imports with flake8
+      run: |
+        flake8 --select=E401 netrd
+        flake8 --select=E401 tests
     - name: Test with pytest
       run: |
         cd tests/
