### 2020-12-17 be4ee2e
actually install flake8
