### 2020-12-17 3a45e7f
restructure __all__s to appease flake8
