+++ netrd/utilities/__init__.py
@@ -10,5 +10,22 @@ from .graph import *
 from .read import *
 from .cluster import *
 from .standardize import *
+from .entropy import *
 
-__all__ = []
+__all__ = [
+    'threshold',
+    'clusterGraph',
+    'js_divergence',
+    'entropy_from_seq',
+    'joint_entropy',
+    'conditional_entropy',
+    'categorized_data',
+    'linear_bins',
+    'create_graph',
+    'undirected',
+    'ensure_undirected',
+    'unweighted',
+    'ensure_unweighted',
+    'read_time_series',
+    'mean_GNP_distance',
+]
