+++ netrd/__init__.py
@@ -9,7 +9,7 @@ Science Insitute 2019 Collabathon.
 
 """
 
-from . import distance
-from . import reconstruction
-from . import dynamics
-from . import utilities
+from . import distance  # noqa
+from . import reconstruction  # noqa
+from . import dynamics  # noqa
+from . import utilities  # noqa
