### 2020-12-17 41a169d
Add flake8 to CI (#320)
* Add flake8 to CI

See #318.

The output of this will currently error:

```
tests/test_dynamics.py:10:1: F401 'numpy as np' imported but unused
netrd/__init__.py:12:1: F401 '.distance' imported but unused
netrd/__init__.py:13:1: F401 '.reconstruction' imported but unused
netrd/__init__.py:14:1: F401 '.dynamics' imported but unused
netrd/__init__.py:15:1: F401 '.utilities' imported but unused
netrd/utilities/__init__.py:8:1: F401 '.threshold.threshold' imported but unused
netrd/utilities/__init__.py:9:1: F401 '.graph.*' imported but unused
netrd/utilities/__init__.py:10
[…]
