+++ netrd/utilities/__init__.py
@@ -6,9 +6,39 @@ Common utilities for use within ``netrd``.
 
 """
 from .threshold import threshold
-from .graph import *
-from .read import *
-from .cluster import *
-from .standardize import *
+from .graph import (
+    create_graph,
+    ensure_undirected,
+    undirected,
+    ensure_unweighted,
+    unweighted,
+)
+from .read import read_time_series
+from .cluster import clusterGraph
+from .standardize import mean_GNP_distance
+from .entropy import (
+    js_divergence,
+    entropy_from_seq,
+    joint_entropy,
+    conditional_entropy,
+    categorized_data,
+    linear_bins,
+)
 
-__all__ = []
+__all__ = [
+    'threshold',
+    'clusterGraph',
+    'js_divergence',
+    'entropy_from_seq',
+    'joint_entropy',
+    'conditional_entropy',
+    'categorized_data',
+    'linear_bins',
+    'create_graph',
+    'undirected',
+    'ensure_undirected',
+    'unweighted',
+    'ensure_unweighted',
+    'read_time_series',
+    'mean_GNP_distance',
+]
