+++ .github/workflows/python-ci.yml
@@ -24,13 +24,17 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install black pytest
+        pip install black pytest flake8
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         pip install -e .
     - name: Lint with black
       run: |
         black --skip-string-normalization --check netrd
         black --skip-string-normalization --check tests
+    - name: Check for unused imports with flake8
+      run: |
+        flake8 --select=F401,F403 netrd
+        flake8 --select=F401,F403 tests
     - name: Test with pytest
       run: |
         cd tests/
