### 2021-01-25 #2802 @arfon
@gabrielspadon @sonal-raj @Shathra @sudarshansudarshan - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission is _netrd: A library for network reconstruction and graph distances_, and the software is primarily written in Python https://github.com/netsiphd/netrd
