### 2021-03-18 #2990 @sonal-raj
@arfon I have reviewed the netrd library along with the supporting paper & documentation and it looks good to me. The checklist has been updated accordingly.
