### 2021-03-22 #2990 @adi3
@sdmccabe in the Usage section of your `README.md`, it seems you are using pseudo-code in the three snippets. Can you please change them to full code examples starting with `import netrd`. Looks like this will be quite straightforward for you as you only need to copy the snippets from [Tutorial](https://netrd.readthedocs.io/en/latest/tutorial.html) documentation.
