### 2021-03-27 #2990 @adi3
@sdmccabe Thanks for the changes. A couple of things minor things in the examples:

- I could not find `time_series_data.txt` and `ground_truth.txt` in your repository. Please state in the documentation where the user can retrieve those files from.
- An alternate approach can be that you have an `/examples` directory which has the 3 python scripts together with the data, and the user simply needs to execute these scripts.
- Let's get rid of the `>>>` in the examples. No need to explicitly state them.

Also, could you please tag me on a comment here once you make any changes? That way I get the notification to proceed. Thanks!
