### 2021-03-29 3380cab
README tweaks
Further tweak the README to follow JOSS reviewer suggestions. Also,
enable syntax highlighting for code examples.
