+++ README.md
@@ -40,16 +40,17 @@ pip install .
 <img src="netrd_reconstruction_example.png" alt="example reconstruction" width="95%"/>
 </p>
 
-The basic usage of a graph reconstruction algorithm is as follows:
+The basic usage of a graph reconstruction algorithm is as follows. Given a file
+of time-series data, `time_series_data.txt`:
 
-```
->>> from netrd.reconstruction import CorrelationMatrix
->>> import numpy as np
->>> TS = np.loadtxt('time_series_data.txt')
->>>
->>> reconstructor = CorrelationMatrix()
->>> G = reconstructor.fit(TS, threshold_type='degree', avg_k=15)
->>> # or alternately, G = reconstructor.results['graph']
+```python
+from netrd.reconstruction import CorrelationMatrix
+import numpy as np
+TS = np.loadtxt('time_series_data.txt')
+
+reconstructor = CorrelationMatrix()
+G = reconstructor.fit(TS, threshold_type='degree', avg_k=15)
+# or alternately, G = reconstructor.results['graph']
 ```
 
 Here, `TS` is an N x L numpy array consisting of L
@@ -76,15 +77,15 @@ details on these parameters.
 
 The basic usage of a distance algorithm is as follows:
 
-```
->>> from netrd.distance import QuantumJSD
->>> import networkx as nx
->>> G1 = nx.fast_gnp_random_graph(1000, .1)
->>> G2 = nx.fast_gnp_random_graph(1000, .1)
->>>
->>> dist_obj = QuantumJSD()
->>> distance = dist_obj.dist(G1, G2)
->>> # or alternatively: distance = dist_obj.results['dist']
+```python
+from netrd.distance import QuantumJSD
+import networkx as nx
+G1 = nx.fast_gnp_random_graph(1000, .1)
+G2 = nx.fast_gnp_random_graph(1000, .1)
+
+dist_obj = QuantumJSD()
+distance = dist_obj.dist(G1, G2)
+# or alternatively: distance = dist_obj.results['dist']
 ```
 
 Here, `G1` and `G2` are `nx.Graph` objects (or subclasses such as
@@ -97,17 +98,18 @@ well as any other values that were computed as a side effect.
 <img src="netrd_dynamics_example.png" alt="example distance" width="95%"/>
 </p>
 
-The basic usage of a dynamics algorithm is as follows:
+The basic usage of a dynamics algorithm is as follows. Given an edgelist,
+`ground_truth.txt`:
 
-```
->>> from netrd.dynamics import VoterModel
->>> import networkx as nx
->>> ground_truth = nx.read_edgelist("ground_truth.txt")
->>>
->>> dynamics_model = VoterModel()
->>> synthetic_TS = dynamics_model.simulate(ground_truth)
->>> # this is the same structure as the input data to a reconstructor
->>> # G = CorrelationMatrix().fit(synthetic_TS)
+```python
+from netrd.dynamics import VoterModel
+import networkx as nx
+ground_truth = nx.read_edgelist("ground_truth.txt")
+
+dynamics_model = VoterModel()
+synthetic_TS = dynamics_model.simulate(ground_truth)
+# this is the same structure as the input data to a reconstructor
+# G = CorrelationMatrix().fit(synthetic_TS)
 ```
 
 This produces a numpy array of time series data.
