### 2021-03-31 #2990 @adi3
@sdmccabe starting to look great! Where are the `time_series_data.txt` and `ground_truth.txt` files? Intuitively, I'd imagine they would be in your `/data` folder, but couldn't find them here.
