### 2021-03-31 #2990 @sdmccabe
They are hypothetical files. I'm not really inclined to make real files for the purposes of illustration; in their place I could also generate simulated data with the correct structure (as in the distances example).
