### 2021-03-31 #2990 @sdmccabe
I think that simulated data would be sufficient to show this, as in the graph distances example. We have more detailed examples in the tutorial and included notebook; simulated data in the README would be sufficient to show the required types of input data.
