### 2021-03-31 6fc37fb
README tweaks (#322)
* Use full examples in README

The [JOSS review](https://github.com/openjournals/joss-reviews/issues/2990)
suggested that we use full examples (including imports) in the README.
This commit doesn't copy/paste the full examples from the tutorial---I
think those benefit from additional context---but does expand out the
existing code snippets to give a clearer picture of how the imports
work, and how the input data are created.

* tweak spacing for consistency between examples

* consistent import orders

* README tweaks

Further tweak the README to follow JOSS reviewer sugge
[…]
