### 2021-03-31 #2990 @adi3
@sdmccabe Your second example runs fine. No success with the first and third though.

- For the first example, the commands run fine but the object `G` does not print anything.
- For the third example, the `simulate` method throws an error:
```
>>> synthetic_TS = dynamics_model.simulate(ground_truth)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/adi/.local/lib/python3.8/site-packages/netrd/utilities/graph.py", line 136, in wrapper
    return func(*args, **kwargs)
TypeError: simulate() missing 1 required positional argument: 'L'
```
