### 2021-03-31 024a435
Correct bug in README example
The dynamics example was missing the number of time steps, a required
parameter, leading to an error when you actually ran it. This commit
fixes that.
