### 2021-03-31 #2990 @sdmccabe
@adi3 the first example returns a NetworkX graph, which does not have a particularly expressive string representation. But you can do something like `print(nx.info(G))` to inspect further.

The third example will be fixed in https://github.com/netsiphd/netrd/pull/323.
