### 2021-05-23 7299498
Fix bug on distance/netsimile.py
