### 2021-05-24 #2990 @adi3
@sdmccabe Thanks for your patience. I went through your paper and the content of it looks quite good. However, it needs to be restructured and condensed a little.

1. Please have two separate sections title `Summary` and `Statement of Need`. Most of the content you present in other sections is either (a) not within the scope of a JOSS publication and should be removed (like the entire `Related software packages` discussion), or (b) can be included in the `Summary` or `Statement of Need` sections. Your `Acknowledgements` section can of course remain as it is.

2. I also would recommend you to cut down on your citations. You have a couple of very lengthy citation blocks, and a paper this short does not warrant 45 citations. I understand that your tool combines a lot of techniques together, so in this case, either cite one or two meta-analysis papers on the topic, or cite only the most important papers.
