### 2021-05-25 #2990 @jkbren
@adi3 no problem, we’re happy that this is progressing, and thanks for the comments.

Is it okay if we give a little pushback on each of them? Structurally, we based much of this off of [another recent JOSS paper about Network Science](https://joss.theoj.org/papers/10.21105/joss.01731), and we were aiming for a paper that was a mix between that style and what’s required in the [JOSS readthedocs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).

- Note 1: Separating the `Summary` from `Statement of Need`: back in October 2020 @arfon made a [similar comment and then amended to say that the paper looked good as-is](https://github.com/openjournals/joss-reviews/issues/2802#issuecomment-719443085). We would be fine separating the two sections, but reading through it we really feel that the Statement of Need is so deeply tied to the Network Science background, that it’s better suited where it is.

- Note 2: Removing the `Related Software Packages` section: we’re happy to do this, but again, we based this off of another JOSS paper that is similar in topic area. Also, in the reviewer checklist there’s a box that says “State of the field: Do the authors describe how this software compares to other commonly-used packages?” and we felt that the section addressed that.

- Note 3: Cutting down on references: Is there a limit we should be aiming for? We're happy to cut down on the references but the motivation doesn't seem very clear - we are implementing a number of methods, all of which are equally deserving of attribution. Given that JOSS is an online journal with no page limits, it seems harmless to include all of them.

We really do appreciate these comments, hoping @arfon can help us figure out how to bring the article in line with JOSS standards.
