### 2021-05-25 #2990 @arfon
Thanks for the thoughtful feedback @adi3 and your response @jkbren. Here are my thoughts:

> - Note 1: Separating the Summary from Statement of Need: back in October 2020 @arfon made a similar comment and then amended to say that the paper looked good as-is. We would be fine separating the two sections, but reading through it we really feel that the Statement of Need is so deeply tied to the Network Science background, that it’s better suited where it is.

I don't have very strong feeling about this except to say that the two seconds following the first, _Network reconstruction from time series data_ and _Comparing networks using graph distances_  read a little like extended versions of the statement of need so I think there may be some merit in trying to refactor the content a little under a separate _Statement of need section_ that describes the core functionality of the package.

> - Note 2: Removing the Related Software Packages section: we’re happy to do this, but again, we based this off of another JOSS paper that is similar in topic area. Also, in the reviewer checklist there’s a box that says “State of the field: Do the authors describe how this software compares to other commonly-used packages?” and we felt that the section addressed that.

I think this I content is fine.

> - Note 3: Cutting down on references: Is there a limit we should be aiming for? We're happy to cut down on the references but the motivation doesn't seem very clear - we are implementing a number of methods, all of which are equally deserving of attribution. Given that JOSS is an online journal with no page limits, it seems harmless to include all of them.

This is also fine as-is. I generally favour liberal citation of the literature (unless it's excessive self-citation 😄).
