### 2021-05-27 #2990 @adi3
@jkbren if you could please work on Note 1 as per @arfon's comments and generate a new PDF, we can move forward with the review. Thank you both!
