+++ paper/paper.md
@@ -37,25 +37,25 @@ date: 29 October 2020
 bibliography: paper.bib
 ---
 
-# Summary and statement of need
+# Statement of need
 
 Complex systems throughout nature and society are often best represented as *networks*. Over the last two decades, alongside the increased availability of large network datasets, we have witnessed the rapid rise of network science [@Amaral2004; @Vespignani2008; @Newman2010; @Barabasi2016]. This field is built around the idea that an increased understanding of the complex structural properties of a variety systems will allow us to better observe, predict, and even control the behavior of these systems.
 
 However, for many systems, the data we have access to is not a direct description of the underlying network. More and more, we see the drive to study networks that have been inferred or reconstructed from non-network data---in particular, using *time series* data from the nodes in a system to infer likely connections between them [@Brugere2018; @Runge2018]. Selecting the most appropriate technique for this task is a challenging problem in network science. Different reconstruction techniques usually have different assumptions, and their performance varies from system to system in the real world. One way around this problem could be to use several different reconstruction techniques and compare the resulting networks. However, network comparison is also not an easy problem, as it is not obvious how best to quantify the differences between two networks, in part because of the diversity of tools for doing so.
 
-The `netrd` Python package seeks to address these two parallel problems in network science by providing, to our knowledge, the most extensive collection of both network reconstruction techniques and network comparison techniques (often referred to as *graph distances*) in a single library. In this article, we detail the two main functionalities of the `netrd` package. Along the way, we describe some of its other useful features. This package builds on commonly used Python packages (e.g. `networkx` [@SciPyProceedings11], `numpy` [@Harris2020], `scipy` [@SciPy2020]) and is already a widely used resource for network scientists and other multidisciplinary researchers. With ongoing open-source development, we see this as a tool that will continue to be used by all sorts of researchers to come.
+The `netrd` Python package seeks to address these two parallel problems in network science.
 
-# Network reconstruction from time series data
+# Summary
 
-Given time series data, $TS$, of the behavior of $N$ nodes / components / sensors of a system over the course of $L$ timesteps, and given the assumption that the behavior of every node, $v_i$, may have been influenced by the past behavior of other nodes, $v_j$, there are dozens of techniques that can be used to infer which connections, $e_{ij}$, are likely to exist between the nodes. That is, we can use one of many *network reconstruction* techniques to create a network representation, $G_r$, that attempts to best capture the relationships between the time series of every node in $TS$. `netrd` is a Python package that lets users perform this network reconstruction task using 17 different techniques, meaning that many different networks can be created from a single time series dataset. For example, in \autoref{fig:ground} we show the outputs of 15 different reconstruction techniques applied to time series data generated from an example network [@Sugihara2012; @Mishchenko2011; @Hoang2019; @Sheikhattar2018; @Friedman2008; @Edelman2005; @Zeng2013; @Donges2009; @Barucca2014; @Ledoit2003; @Stetter2012; @Peixoto2019].
+`netrd` offers, to our knowledge, the most extensive collection of both network reconstruction techniques and network comparison techniques (often referred to as *graph distances*) in a single library. Below, we expand on these two main functionalities of the `netrd` package.
 
-## Simulated network dynamics
[…]
