+++ paper/paper.md
@@ -53,7 +53,7 @@ The first core use of `netrd` is to reconstruct networks from time series data. 
 
 Practitioners often apply these network reconstruction algorithms to real time series data. For example, in neuroscience, researchers often try to reconstruct functional networks from time series readouts of neural activity [@Mishchenko2011]. In economics, researchers can infer networks of influence between companies based on time series of changes in companies' stock prices [@Squartini2018]. At the same time, it is often quite helpful having the freedom to *simulate* arbitrary time series dynamics on randomly generated networks. This provides a controlled setting to assess the performance of network reconstruction algorithms. For this reason, the `netrd` package also includes a number of different techniques for simulating dynamics on networks.
 
-The second core use of `netrd` addresses a common goal when studying networks: describing and quantifying the differences between two networks. This is a challenging problem, as there are countless axes upon which two networks can differ; as such, a number of *graph distance* measures have emerged over the years attempting to address this problem. As is the case for many problems in network science, it can be difficult to know which (of many) measures are suited for a given setting. In `netrd`, we consolidate over 20 different graph distance measures into a single package [@Jaccard1901; @Hamming1950; @Jurman2015; @Golub2013; @Donnat2018; @Carpi2011; @Bagrow2019; @DeDomenico2016; @Chen2018; @Hammond2013; @Monnig2018; @Tsitsulin2018; @Jurman2011; @Ipsen2002; @Torres2019; @Mellor2019; @Schieber2017; @Koutra2016; @Berlingerio2012]. \autoref{fig:dists} shows an example of just how different these measures can be when comparing two networks, $G_1$ and $G_2$. This submodule in `netrd` has already been used in recent work with a novel characterization of the graph distance literature [@Hartle2020].
+The second core use of `netrd` addresses a common goal when studying networks: describing and quantifying the differences between two networks. This is a challenging problem, as there are countless axes upon which two networks can differ; as such, a number of *graph distance* measures have emerged over the years attempting to address this problem. As is the case for many hard problems in network science, it can be difficult to know which (of many) measures are suited for a given setting. In `netrd`, we consolidate over 20 different graph distance measures into a single package [@Jaccard1901; @Hamming1950; @Jurman2015; @Golub2013; @Donnat2018; @Carpi2011; @Bagrow2019; @DeDomenico2016; @Chen2018; @Hammond2013; @Monnig2018; @Tsitsulin2018; @Jurman2011; @Ipsen2002; @Torres2019; @Mellor2019; @Schieber2017; @Koutra2016; @Berlingerio2012]. \autoref{fig:dists} shows an example of just how different these measures can be when comparing two networks, $G_1$ and $G_2$. This submodule in `netrd` has already been used in recent work with a novel characterization of the graph distance literature [@Hartle2020].
 
 This package builds on commonly used Python packages (e.g. `networkx` [@SciPyProceedings11], `numpy` [@Harris2020], `scipy` [@SciPy2020]) and is already a widely used resource for network scientists and other multidisciplinary researchers. With ongoing open-source development, we see this as a tool that will continue to be used by all sorts of researchers to come.
 
