### 2021-05-27 423b434
paper edits (and added a reference to the README, as well as the JOSS pdf generator workflow) (#325)
* paper edits

* add pretty plots to the readme

* tweak readme

* paper edits

* Create draft-pdf.yml

* Update draft-pdf.yml

* Update draft-pdf.yml

* try pdf compiler action

* try pdf compiler action

* update readme

* update readme

Co-authored-by: jkbren <klein.br@husky.neu.edu>
