### 2021-05-28 #2990 @arfon
@jkbren @sdmccabe – thanks for your updates to the paper. This is looking good!

At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
