### 2021-06-01 1448556
Bump version to 0.3
This PR bumps the version to align with the end of our [JOSS
reviews](https://github.com/openjournals/joss-reviews/issues/2990).
