### 2021-06-01 f703c19
Bump version to 0.3 (#326)
This PR bumps the version to align with the end of our [JOSS
reviews](https://github.com/openjournals/joss-reviews/issues/2990).
